requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
firebase-admin==4.5.3
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httplib2==0.19.1
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==5.1.0
packaging==20.9
priority==1.3.0
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
Quart==0.14.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
wsproto==1.0.0
yarl==1.6.3